Better support module-aware mode context
Currently dockexec runs with /pwd mounted as the directory of the
package under test. However, this means that for a subpackage in a
module we lose the module context because the parent directories are not
mounted. This causes problems for any tests that run commands, e.g.
cmd/go, that would otherwise assume we are still in a module context.

Instead, assuming:

* a module $m is rooted at $moddir
* that the package $m/cmd/blah/ exists
* a working directory of $moddir
* that we run go test -exec='...' ./cmd/blah

Then $moddir will be mounted as /start and the working directory will be
/start/cmd/blah.

We also mount GOPATH and GOCACHE at canonical locations (see the comment
for buildDockerFlags()).

Per the TODO comment, we have not implemented the handling of GOPATH
mode. Instead, if we are not in module-aware mode with a main module we
fall back to the logic of ad hoc mode.