main: handle dockexec being called with package that is module root
Given a module called foo.com, and a package at the root of that module,
the result binary name is foo.com. Improve the regex that is used to
find the Go binary in the arguments passed to dockexec to support such
cases. Add a test case to cover it.