update linux images used by the tests
Recent versions of Go build with recent versions of Glibc.
And because some of the tests mount the host's GOROOT,
and then call it via "go env",
the guest must also have a similarly recent glibc version.

This fails at the moment on Arch with glibc 2.34,
since the busybox glibc image bundles glibc from debian stable,
which at the moment is 2.31:

	> exec go test -exec='dockexec '$IMAGE' -v '$HOST_GOROOT':/goroot ' -v main_test.go
	[stdout]
	=== RUN   TestOutput
	working directory: /start
		main_test.go:29: failed to run /goroot/bin/go env -json: exit status 1
			/goroot/bin/go: /lib/libc.so.6: version `GLIBC_2.32' not found (required by /goroot/bin/go)
	--- FAIL: TestOutput (0.00s)

Switch to a pinned debian testing image, which uses glibc 2.33.

For the other busybox images, bump to the latest version available.