remove explicit support for GOPATH mode
This includes mounting GOPATH directories as well.
When in module mode, GOPATH is only used as a fallback location
to place other directories under, such as GOBIN and GOMODCACHE.

If we feel like dockexec needs to expose those via volume mounts,
then we can do so just like we do for GOCACHE.

We share GOCACHE because it makes sense for any tests that themselves
use the Go toolchain; a temporary GOCACHE can waste resources.

For the same reason, share GOMODCACHE, which was introduced in go env in
early 2020. We were already often sharing it via GOPATH, as its default
location is GOPATH/pkg/mod, so this prevents a regression.
Not sharing GOMODCACHE can trigger downloading modules again and again.

The only missing bit from GOPATH that we stop sharing is GOBIN,
as its default location is GOPATH/bin.
This seems fine; we don't share other binary directories from the host
like /usr/bin, and we don't want the guest to save binaries in the host.