update deps, swap docker-compose for docker compose
The latter allows us to clean up a number of TODOs.
The new docker compose, the v2 of docker-compose, came out in 2020
and the old docker-compose v1 stopped receiving updates in 2023
nor is it shipped as part of Docker any longer.
On my system, docker-compose is now an alias for docker compose.

For those reasons, we can now assume v2 is available.